Refactor, rename and generally clean-up codebase
- rename package_dir --> package_path
- remove unused imports from builder, utils, pr_list.__init__,
- move cache from builder to pr_list
- move status_file from utils to pr_list
- move status_from_cache from server to pr_list
- change outdated_pr_list --> update_pr_list
