+++ procbuild/builder.py
@@ -11,19 +11,9 @@ import random
 from os.path import join as joinp
 from glob import iglob
 
+from . import package_path
+
 excluded = ['vanderwalt', '00_vanderwalt', 'jane_doe', 'bibderwalt', '00_intro']
-
-base_path = os.path.abspath(os.path.dirname(__file__))
-
-
-def cache(path='../cache'):
-    cache_path = joinp(base_path, path)
-    try:
-        os.mkdir(cache_path)
-    except OSError as e:
-        pass
-
-    return cache_path
 
 
 def repo(user='scipy'):
@@ -121,7 +111,7 @@ class BuildManager:
         data_filenames = ['IEEEtran.cls', 
                           'draftwatermark.sty', 
                           'everypage.sty']
-        self.data_files = [joinp(base_path, 'data', f) for f in data_filenames]
+        self.data_files = [joinp(package_path, 'data', f) for f in data_filenames]
 
     def add_output(self, msg):
         self.build_output += msg
