+++ procbuild/utils.py
@@ -3,15 +3,15 @@ import inspect
 import io
 import time
 
-from datetime import datetime, timedelta
+from datetime import datetime
 from os.path import join as joinp
 
-from . import package_dir
-from .builder import cache
+from . import package_path
+
 
 def log(message):
     print(message)
-    with io.open(joinp(package_dir, '../flask.log'), 'a') as f:
+    with io.open(joinp(package_path, '../flask.log'), 'a') as f:
         time_of_message = time.strftime("%Y-%m-%d %H:%M:%S", time.gmtime()) 
         cf = inspect.currentframe().f_back
         where = f'{cf.f_code.co_filename}:{cf.f_lineno}'
@@ -32,6 +32,4 @@ def file_age(fn):
     return delta.seconds / 60
 
     
-def status_file(nr):
-    return joinp(cache(), str(nr) + '.status')
